+++ kong/tools/queue.lua
@@ -82,7 +82,6 @@ local semaphore_new = semaphore.new
 local math_min = math.min
 local now = ngx.now
 local sleep = ngx.sleep
-local worker_exiting = ngx.worker.exiting
 local null = ngx.null
 
 
@@ -242,8 +241,12 @@ function Queue:process_once()
 
   -- We've got our first entry from the queue.  Collect more entries until max_coalescing_delay expires or we've collected
   -- max_batch_size entries to send
+  if ngx.worker.exiting() then
+    -- minimize coalescing delay during shutdown to quickly process remaining entries
+    self.max_coalescing_delay = COALESCE_MIN_TIME
+  end
   while entry_count < self.max_batch_size
-    and self.max_coalescing_delay - (now() - data_started) >= COALESCE_MIN_TIME and not worker_exiting()
+    and self.max_coalescing_delay - (now() - data_started) >= COALESCE_MIN_TIME
   do
     -- Instead of waiting for the coalesce time to expire, we cap the semaphore wait to COALESCE_POLL_TIME
     -- so that we can check for worker shutdown periodically.