+++ spec/01-unit/27-queue_spec.lua
@@ -94,7 +94,12 @@ describe("plugin queue", function()
           else
             return real_now()
           end
-        end
+        end,
+        worker = {
+          exiting = function()
+            return false
+          end
+        }
       }
     })
   end)
@@ -206,6 +211,40 @@ describe("plugin queue", function()
   end)
 
   it("batches messages when `max_batch_size` is 2", function()
+    local process_count = 0
+    local first_entry, last_entry
+    local function enqueue(entry)
+      Queue.enqueue(
+        queue_conf({
+          name = "batch",
+          max_batch_size = 2,
+          max_coalescing_delay = 0.1,
+        }),
+        function(_, batch)
+          first_entry = first_entry or batch[1]
+          last_entry = batch[#batch]
+          process_count = process_count + 1
+          return true
+        end,
+        nil,
+        entry
+      )
+    end
+    enqueue("One")
+    enqueue("Two")
+    enqueue("Three")
+    enqueue("Four")
+    enqueue("Five")
+    wait_until_queue_done("batch")
+    assert.equals(3, process_count)
+    assert.equals("One", first_entry)
+    assert.equals("Five", last_entry)
+  end)
+
+  it("batches messages during shutdown", function()
+    _G.ngx.worker.exiting = function()
+      return true
+    end
     local process_count = 0
     local first_entry, last_entry
     local function enqueue(entry)